Move listFileSlices to HoodieBaseRelation
